Converted to float hash codes

+++ src/Foundation/Center.php
@@ -34,7 +34,7 @@ class Center extends BuildWidget {
 		);
 	}
 
-	public function getHashCode(): int {
+	public function getHashCode(): float {
 		return $this->child->getHashCode();
 	}
 }
