+++ src/Hashable.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 namespace Elephox\Templar;
 
 interface Hashable {
-	public function getHashCode(): int;
+	public function getHashCode(): float;
 }
